Version 1.3
Now checks if the video element has been reloaded. This is done by checking if the script is running for the first time and a paused video is detected. This feature will prevent the page from from randomly
pausing the video outside of the nags and page reloads.  Also wont run the script on pages that dont have "watch" in the url to prevent runtime on the search and index pages of youtube.